doc2txt_api/pyproject.toml
[project]
name = "doc2txt_api"
version = "0.1"
dependencies = [
    "beautifulsoup4==4.12.3",
    "Flask==3.0.3",
    "gunicorn==22.0.0",
    "pydantic==2.7.4",
    "PyJWT==2.8.0",
    "pytesseract==0.3.10",
    "pdfminer.six==20240706",
    "pdf2image==1.17.0",
    "Wand==0.6.13",
]